+++ apps/fluent-tester/package.json
@@ -24,6 +24,7 @@
     "bundle": "fluentui-scripts bundle"
   },
   "dependencies": {
+    "@fluentui-react-native/apple-theme": ">=0.1.0 <1.0.0",
     "@fluentui-react-native/default-theme": ">=0.4.1 <1.0.0",
     "@fluentui-react-native/experimental-avatar": ">=0.4.0 <1.0.0",
     "@fluentui-react-native/experimental-button": "0.4.0",
