+++ apps/fluent-tester/src/FluentTester/theme/CustomThemes.ts
@@ -1,10 +1,14 @@
 import { createDefaultTheme, ThemeOptions } from '@fluentui-react-native/default-theme';
+import { createAppleTheme } from '@fluentui-react-native/apple-theme';
 import { ThemeReference } from '@fluentui-react-native/theme';
 import { applyBrand, OfficeBrand } from './applyBrand';
 import { applyTheme, ThemeNames } from './applyTheme';
+import { Platform } from 'react-native';
 
 const themeOptions: ThemeOptions = { paletteName: 'TaskPane', appearance: 'dynamic' };
-const baseTheme = createDefaultTheme(themeOptions);
+
+const isApplePlatform = Platform.OS === 'macos' || Platform.OS === 'ios';
+const baseTheme = isApplePlatform ? createAppleTheme(themeOptions) : createDefaultTheme(themeOptions);
 
 export const lightnessOptions = [
   { label: 'Auto', value: 'dynamic' },
