+++ packages/theming/apple-theme/src/index.ts
@@ -1,4 +1,2 @@
+export { defaultAppleTheme, defaultAppleDarkTheme } from './appleTheme';
 export * from './createAppleTheme';
-export * from './createPartialAppleTheme';
-export * from './NativeModule/index';
-export * from './paletteFromAppleColors';
